1-last_digit.c

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -9,7 +9,7 @@
  */
 int main(void)
 {
-	long int  n, rem;
+	int  n, rem;
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
